Add necessary parts of framework written by F.H.
This code is copyrighted by F.H. but I have express permission to publish these parts. For the whole thing to run you'll need specific setup instructions regarding naoQI, python 2.7, and some other dependencies as well as wrappers to simplify the API. All of this is included in the package/repo F.H. provided to the students starting with the project.