Implement I spy
It works just fine even though I struggled quite a bit with the naoQI library and python itself during development. There's certainly a bunch of improvements that could be made like better naming, more configurability, reuse of scans between games to decrease wait time, etc. but all in all it does what it was supposed to quite well and got a max grade.